Move some type definitions back into the public API.
Since none of the internal code still relies on these type definitions, remove
the `D.L.I.Normal` and `D.L.I.Monoidal` modules, and move the types to the
`D.L.Normal` and `D.L.Monoidal` modules.
